Exercise 5.42: Implement letcc and throw
"letcc Id in Expr" can be used to capture a continuation anywhere.
Such a captured continuation may be invoked with "throw".

The expression "throw Expr to Expr" evaluates the two subexpressions.
The second expression should return a continuation, which is applied
to the value of the first expression. The current continuation of
the throw expression is ignored.